we can plot the waterlevels using dates
The problem is that, while it seems the dates are sorted (so it must
have some kind of sense of date sort order, or maybe it does a numeric
sort on the strings? the latter works because the format is
year/month/day 24-hour:minutes), they are not spaced properly, meaning
two dates separated by a week or a minute will be equally close on the x
axis. I wanted to use the dates converted to seconds on the x axis so
they'd be spaced correctly, but I had trouble then formatting the date
in javascript from the seconds.  Anyhow we'll leave it like this for
now.